requirements.txt
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.6
cryptography==35.0.0
flake8==3.9.2
mccabe==0.6.1
paramiko==2.8.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyNaCl==1.4.0
python-dotenv==0.19.1
six==1.16.0
toml==0.10.2
